test: globally install tar instead of linking
